+++ sample/privatezone/example_privatezone_bind_vpc.go
@@ -2,7 +2,6 @@ package privatezonesample
 
 import (
 	"fmt"
-	"github.com/baidubce/baiducloud-go-sdk/bce"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/privatezone"
 )
@@ -17,7 +16,6 @@ func BindVpc() {
 	}
 	bindVpcRequest := &privatezone.BindVpcRequest{
 		ZoneId:      util.PtrString(""),
-		Action:      util.PtrString(""),
 		ClientToken: util.PtrString(""),
 		Region:      util.PtrString(""),
 		VpcIds:      []*string{},
@@ -25,13 +23,6 @@ func BindVpc() {
 	err = client.BindVpc(bindVpcRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
-		if serviceErr, ok := err.(*bce.BceServiceError); ok {
-			fmt.Printf("request failed, RequestId: %s, Code: %s, Message: %s\n",
-				serviceErr.RequestId, serviceErr.Code, serviceErr.Message)
-		} else {
-			fmt.Println("request failed:", err)
-		}
-		return
+		fmt.Println("request failed:", err)
 	}
-	fmt.Println("BindVpc success")
 }