+++ sample/privatezone/example_privatezone_enable_record.go
@@ -1,7 +1,6 @@
 package privatezonesample
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/privatezone"
@@ -20,16 +19,9 @@ func EnableRecord() {
 		Action:      util.PtrString(""),
 		ClientToken: util.PtrString(""),
 	}
-	result, err := client.EnableRecord(enableRecordRequest)
+	err = client.EnableRecord(enableRecordRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
 		fmt.Println("request failed:", err)
-		return
 	}
-	data, err := json.MarshalIndent(result, "", "    ")
-	if err != nil {
-		fmt.Println("json marshalIndent failed:", err)
-		return
-	}
-	fmt.Println(string(data))
 }