+++ sample/privatezone/example_privatezone_unbind_vpc.go
@@ -2,7 +2,6 @@ package privatezonesample
 
 import (
 	"fmt"
-	"github.com/baidubce/baiducloud-go-sdk/bce"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/privatezone"
 )
@@ -19,19 +18,12 @@ func UnbindVpc() {
 		ZoneId:      util.PtrString(""),
 		Action:      util.PtrString(""),
 		ClientToken: util.PtrString(""),
-		Region:      util.PtrString("bj"),
-		VpcIds:      []*string{util.PtrString("vpc-8zn7k6fny75x")},
+		Region:      util.PtrString(""),
+		VpcIds:      []*string{},
 	}
 	err = client.UnbindVpc(unbindVpcRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
-		if serviceErr, ok := err.(*bce.BceServiceError); ok {
-			fmt.Printf("request failed, RequestId: %s, Code: %s, Message: %s\n",
-				serviceErr.RequestId, serviceErr.Code, serviceErr.Message)
-		} else {
-			fmt.Println("request failed:", err)
-		}
-		return
+		fmt.Println("request failed:", err)
 	}
-	fmt.Println("UnbindVpc success")
 }