+++ services/privatezone/model_bind_vpc_request.go
@@ -2,7 +2,6 @@ package privatezone
 
 type BindVpcRequest struct {
 	ZoneId      *string   `json:"-"`
-	Action      *string   `json:"-"`
 	ClientToken *string   `json:"-"`
 	Region      *string   `json:"region,omitempty"`
 	VpcIds      []*string `json:"vpcIds,omitempty"`