+++ services/privatezone/model_disable_record_request.go
@@ -3,5 +3,4 @@ package privatezone
 type DisableRecordRequest struct {
 	RecordId    *string `json:"-"`
 	ClientToken *string `json:"-"`
-	Action      *string `json:"-"`
 }